* next: (116 commits)
  Adding new example script for using the new PFB arbitrary resampler interface. One resampler takes user-generated taps and another resampler just takes the resampling rate. Both input and output signals are plotted.
  Modifying blsk2 wrapper for PFB arbitrary resampler to allow the user to just specify the requested resampling rate without providing their own filter taps.
  uhd: reverting tag changes on uhd single usrp source, there seems to be issues with the work() logic
  uhd: removed default value chan=0 in the cc files
  WITH_INCLUDES _must_ be last
  uhd: replaced CFLAGS with CPPFLAGS variable, and fixed swig args FIXME
  uhd: default channel params to zero for single source and sink blocks
  Revert "Removed usrp2-firmware from being automatically built."
  Remove generated file from repo
  Removed usrp2-firmware from being automatically built.
  uhd: added libdir to UHD CFLAGS (shared by lib and swig)
  Swapping out preset keys until I work out some of their issues.
  Modifying QA tests for the sample tags. By default, it only checks the sizes of the tags since order is not specified or guarenteed.
  Block is a gr_block, so this sets its relative rate. Was required for using in the QA of the sample tags code.
  Removing global pmt constants. Were causing segfaults during make check. Must fix this later.
  Fixing up the UHD sample tag example to take command line options.
  Swapping order of testing rrate.
  Changing propagation policy enum type name and making a few other minor edits.
  Changing API for gr_skiphead to use uint64_t for the offset instead of size_t (still unsigned). Fixes issue #304.
  Adding typedef for uint64_t and int64_t so we can use them through SWIG.
  ...

Passes make distcheck.

+++ gnuradio-core/src/lib/general/gr_float_to_complex.cc
@@ -52,12 +52,12 @@ gr_float_to_complex::work (int noutput_items,
 
   switch (input_items.size ()){
   case 1:
-    for (int j = 0; j < noutput_items*d_vlen; j++)
+    for (size_t j = 0; j < noutput_items*d_vlen; j++)
       out[j] = gr_complex (r[j], 0);
     break;
 
   case 2:
-    for (int j = 0; j < noutput_items*d_vlen; j++)
+    for (size_t j = 0; j < noutput_items*d_vlen; j++)
       out[j] = gr_complex (r[j], i[j]);
     break;
